Removed redundant Auth0 Viewcontroller + Fixes in UIViewControlerLocalGlobal
The content of Auth0ViewController was already handled in UIViewControllerLocalGlobal, so it should have been removed. The identity token is not used and could be removed. I simplified the button delegates and tried to fix the indentation, but it was correctly displayed in Visual studio. I removed unused methods, as we can assume that a token remains valid (as far as I know).